+++ source/isaaclab/isaaclab/controllers/utils.py
@@ -14,6 +14,7 @@ from isaacsim.core.utils.extensions import enable_extension
 
 enable_extension("isaacsim.asset.exporter.urdf")
 
+import nvidia.srl.tools.logger as logger
 import omni.log
 from nvidia.srl.from_usd.to_urdf import UsdToUrdf
 
@@ -33,6 +34,7 @@ def convert_usd_to_urdf(usd_path: str, output_path: str, force_conversion: bool 
         "edge_names_to_remove": None,
         "root": None,
         "parent_link_is_body_1": None,
+        "log_level": logger.level_from_name("ERROR"),
     }
 
     urdf_output_dir = os.path.join(output_path, "urdf")