+++ source/isaaclab/docs/CHANGELOG.rst
@@ -1,7 +1,7 @@
 Changelog
 ---------
 
-0.36.14 (2025-04-09)
+0.36.15 (2025-04-09)
 ~~~~~~~~~~~~~~~~~~~~
 
 Changed
@@ -12,7 +12,7 @@ Changed
   the cuda device, which results in NCCL errors on distributed setups.
 
 
-0.36.13 (2025-04-01)
+0.36.14 (2025-04-01)
 ~~~~~~~~~~~~~~~~~~~~
 
 Fixed
@@ -21,7 +21,7 @@ Fixed
 * Added check in RecorderManager to ensure that the success indicator is only set if the termination manager is present.
 
 
-0.36.12 (2025-03-24)
+0.36.13 (2025-03-24)
 ~~~~~~~~~~~~~~~~~~~~
 
 Changed
@@ -29,6 +29,15 @@ Changed
 
 * Definition of render settings in :class:`~isaaclab.sim.SimulationCfg` is changed to None, which means that
   the default settings will be used from the experience files and the double definition is removed.
+
+
+0.36.12 (2025-03-19)
+~~~~~~~~~~~~~~~~~~~~
+
+Added
+^^^^^
+
+* Added parameter to show warning if Pink IK solver fails to find a solution.
 
 
 0.36.11 (2025-03-19)